-rearrange due to clearer module hierarchy -introduce accessor.py and core.py for the sake of the first point
